Pull up Report namespace
