+++ lib/reek/spec/should_reek_only_of.rb
@@ -1,5 +1,5 @@
 require_relative '../examiner'
-require_relative '../cli/report/formatter'
+require_relative '../report/formatter'
 
 module Reek
   module Spec
@@ -21,7 +21,7 @@ module Reek
       end
 
       def failure_message
-        rpt = CLI::Report::Formatter.format_list(@warnings)
+        rpt = Report::Formatter.format_list(@warnings)
         "Expected #{@examiner.description} to reek only of #{@smell_category}, but got:\n#{rpt}"
       end
