+++ spec/reek/cli/option_interpreter_spec.rb
@@ -1,6 +1,7 @@
 require_relative '../../spec_helper'
 
 require_relative '../../../lib/reek/cli/options'
+require_relative '../../../lib/reek/report/report'
 
 RSpec.describe Reek::CLI::OptionInterpreter do
   let(:options) { OpenStruct.new }
@@ -8,7 +9,7 @@ RSpec.describe Reek::CLI::OptionInterpreter do
 
   describe '#reporter' do
     it 'returns a Report::TextReport instance by default' do
-      expect(instance.reporter).to be_instance_of Reek::CLI::Report::TextReport
+      expect(instance.reporter).to be_instance_of Reek::Report::TextReport
     end
   end
 end
