+++ spec/reek/report/html_report_spec.rb
@@ -1,9 +1,9 @@
 require_relative '../../spec_helper'
 require_relative '../../../lib/reek/examiner'
-require_relative '../../../lib/reek/cli/report/report'
+require_relative '../../../lib/reek/report/report'
 
-RSpec.describe Reek::CLI::Report::HTMLReport do
-  let(:instance) { Reek::CLI::Report::HTMLReport.new }
+RSpec.describe Reek::Report::HTMLReport do
+  let(:instance) { Reek::Report::HTMLReport.new }
 
   context 'with an empty source' do
     let(:examiner) { Reek::Examiner.new('') }
