+++ spec/reek/report/xml_report_spec.rb
@@ -1,10 +1,10 @@
 require_relative '../../spec_helper'
 require_relative '../../../lib/reek/examiner'
-require_relative '../../../lib/reek/cli/report/report'
-require_relative '../../../lib/reek/cli/report/formatter'
+require_relative '../../../lib/reek/report/report'
+require_relative '../../../lib/reek/report/formatter'
 
-RSpec.describe Reek::CLI::Report::XMLReport do
-  let(:instance) { Reek::CLI::Report::XMLReport.new }
+RSpec.describe Reek::Report::XMLReport do
+  let(:instance) { Reek::Report::XMLReport.new }
 
   context 'empty source' do
     let(:examiner) { Reek::Examiner.new('') }
